+++ taf/auth_repo.py
@@ -1,13 +1,21 @@
 import json
 import os
+import tempfile
 from collections import defaultdict
+from contextlib import contextmanager
 from pathlib import Path
 from subprocess import CalledProcessError
+from tuf.repository_tool import METADATA_DIRECTORY_NAME
 from taf.log import taf_logger
 from taf.git import GitRepository, NamedGitRepository
+from taf.repository_tool import (
+    Repository as TAFRepository,
+    get_role_metadata_path,
+    get_target_path,
+)
 
 
-class AuthRepoMixin(object):
+class AuthRepoMixin(TAFRepository):
 
     LAST_VALIDATED_FILENAME = "last_validated_commit"
     TEST_REPO_FLAG_FILE = "test-auth-repo"
@@ -21,14 +29,14 @@ class AuthRepoMixin(object):
         """
         # the repository's name consists of the namespace and name (namespace/name)
         # the configuration directory should be _name
-        last_dir = os.path.basename(os.path.normpath(self.repo_path))
-        conf_path = Path(self.repo_path).parent / f"_{last_dir}"
+        last_dir = os.path.basename(os.path.normpath(self.path))
+        conf_path = Path(self.path).parent / f"_{last_dir}"
         conf_path.mkdir(parents=True, exist_ok=True)
         return str(conf_path)
 
     @property
     def certs_dir(self):
-        certs_dir = Path(self.repo_path, "certs")
+        certs_dir = Path(self.path, "certs")
         certs_dir.mkdir(parents=True, exist_ok=True)
         return str(certs_dir)
 
@@ -45,7 +53,7 @@ class AuthRepoMixin(object):
     def get_target(self, target_name, commit=None, safely=True):
         if commit is None:
             commit = self.head_commit_sha()
-        target_path = (Path(self.targets_path) / target_name).as_posix()
+        target_path = get_target_path(target_name)
         if safely:
             return self._safely_get_json(commit, target_path)
         else:
@@ -63,12 +71,37 @@ class AuthRepoMixin(object):
                 continue
         return False
 
+    @contextmanager
+    def repository_at_revision(self, commit):
+        """
+        Context manager which makes sure that TUF repository is instantiated
+        using metadata files at the specified revision. Creates a temp directory
+        and metadata files inside it. Deleted the temp directory when no longer
+        needed.
+        """
+        tuf_repository = self._tuf_repository
+        with tempfile.TemporaryDirectory() as temp_dir:
+            metadata_files = self.list_files_at_revision(
+                commit, METADATA_DIRECTORY_NAME
+            )
+            Path(temp_dir, METADATA_DIRECTORY_NAME).mkdir(parents=True)
+            for file_name in metadata_files:
+                path = Path(temp_dir, METADATA_DIRECTORY_NAME, file_name)
+                with open(path, "w") as f:
+                    data = self.get_json(
+                        commit, f"{METADATA_DIRECTORY_NAME}/{file_name}"
+                    )
+                    json.dump(data, f)
+            self._load_tuf_repository(temp_dir)
+            yield
+            self._tuf_repository = tuf_repository
+
     def set_last_validated_commit(self, commit):
         """
         Set the last validated commit of the authentication repository
         """
         taf_logger.debug(
-            "Auth repo {}: setting last validated commit to: {}", self.repo_name, commit
+            "Auth repo {}: setting last validated commit to: {}", self.name, commit
         )
         Path(self.conf_dir, self.LAST_VALIDATED_FILENAME).write_text(commit)
 
@@ -102,47 +135,54 @@ class AuthRepoMixin(object):
                 previous_commits[target_path] = target_commit
         taf_logger.debug(
             "Auth repo {}: new commits per repositories according to targets.json: {}",
-            self.repo_name,
+            self.name,
             repositories_commits,
         )
         return repositories_commits
 
     def target_commits_at_revisions(self, commits):
         targets = defaultdict(dict)
-        for commit in commits:
-            targets_at_revision = self._safely_get_json(
-                commit, self.metadata_path + "/targets.json"
-            )
-            if targets_at_revision is None:
-                continue
-            targets_at_revision = targets_at_revision["signed"]["targets"]
 
+        for commit in commits:
+            # repositories.json might not exit, if the current commit is
+            # the initial commit
             repositories_at_revision = self._safely_get_json(
-                commit, self.targets_path + "/repositories.json"
+                commit, get_target_path("repositories.json")
             )
             if repositories_at_revision is None:
                 continue
             repositories_at_revision = repositories_at_revision["repositories"]
 
-            for target_path in targets_at_revision:
-                if target_path not in repositories_at_revision:
-                    # we only care about repositories
-                    continue
-                try:
-                    target_content = self.get_json(
-                        commit, self.targets_path + "/" + target_path
-                    )
-                    target_commit = target_content.get("commit")
-                    target_branch = target_content.get("branch", "master")
-                    targets[commit][target_path] = (target_branch, target_commit)
-                except json.decoder.JSONDecodeError:
-                    taf_logger.debug(
-                        "Auth repo {}: target file {} is not a valid json at revision {}",
-                        self.repo_name,
-                        target_path,
-                        commit,
-                    )
-                    continue
+            # get names of all targets roles defined in the current revision
+            with self.repository_at_revision(commit):
+                roles_at_revision = self.get_all_targets_roles()
+
+            for role_name in roles_at_revision:
+                # targets metadata files corresponding to the found roles must exist
+                targets_at_revision = self.get_json(
+                    commit, get_role_metadata_path(role_name)
+                )
+                targets_at_revision = targets_at_revision["signed"]["targets"]
+
+                for target_path in targets_at_revision:
+                    if target_path not in repositories_at_revision:
+                        # we only care about repositories
+                        continue
+                    try:
+                        target_content = self.get_json(
+                            commit, get_target_path(target_path)
+                        )
+                        target_commit = target_content.get("commit")
+                        target_branch = target_content.get("branch", "master")
+                        targets[commit][target_path] = (target_branch, target_commit)
+                    except json.decoder.JSONDecodeError:
+                        taf_logger.debug(
+                            "Auth repo {}: target file {} is not a valid json at revision {}",
+                            self.name,
+                            target_path,
+                            commit,
+                        )
+                        continue
         return targets
 
     def _safely_get_json(self, commit, path):
@@ -151,48 +191,39 @@ class AuthRepoMixin(object):
         except CalledProcessError:
             taf_logger.info(
                 "Auth repo {}: {} not available at revision {}",
-                self.repo_name,
+                self.name,
                 os.path.basename(path),
                 commit,
             )
         except json.decoder.JSONDecodeError:
             taf_logger.info(
                 "Auth repo {}: {} not a valid json at revision {}",
-                self.repo_name,
+                self.name,
                 os.path.basename(path),
                 commit,
             )
 
 
-class AuthenticationRepo(AuthRepoMixin, GitRepository):
+class AuthenticationRepo(GitRepository, AuthRepoMixin):
     def __init__(
-        self,
-        repo_path,
-        metadata_path="metadata",
-        targets_path="targets",
-        repo_urls=None,
-        additional_info=None,
-        default_branch="master",
+        self, path, repo_urls=None, additional_info=None, default_branch="master"
     ):
-        super().__init__(repo_path, repo_urls, additional_info, default_branch)
-        self.targets_path = targets_path
-        self.metadata_path = metadata_path
+        super().__init__(path, repo_urls, additional_info, default_branch)
 
 
-class NamedAuthenticationRepo(AuthRepoMixin, NamedGitRepository):
+class NamedAuthenticationRepo(NamedGitRepository, AuthRepoMixin):
     def __init__(
         self,
         root_dir,
         repo_name,
-        metadata_path="metadata",
-        targets_path="targets",
         repo_urls=None,
         additional_info=None,
         default_branch="master",
     ):
-
         super().__init__(
-            root_dir, repo_name, repo_urls, additional_info, default_branch
+            root_dir=root_dir,
+            repo_name=repo_name,
+            repo_urls=repo_urls,
+            additional_info=additional_info,
+            default_branch=default_branch,
         )
-        self.targets_path = targets_path
-        self.metadata_path = metadata_path