+++ taf/developer_tool.py
@@ -282,7 +282,7 @@ def create_repository(
             return
 
     tuf.repository_tool.METADATA_STAGED_DIRECTORY_NAME = METADATA_DIRECTORY_NAME
-    repository = create_new_repository(auth_repo.repo_path)
+    repository = create_new_repository(auth_repo.path)
 
     def _sort_roles(key_info, repository):
         # load keys not stored on YubiKeys first, to avoid entering pins
@@ -348,8 +348,7 @@ def create_repository(
     # if the repository is a test repository, add a target file called test-auth-repo
     if test:
         test_auth_file = (
-            Path(auth_repo.repo_path, auth_repo.targets_path)
-            / auth_repo.TEST_REPO_FLAG_FILE
+            Path(auth_repo.path, auth_repo.targets_path) / auth_repo.TEST_REPO_FLAG_FILE
         )
         test_auth_file.touch()
         targets_obj = _role_obj("targets", repository)
@@ -691,9 +690,9 @@ def generate_repositories_json(
         url = target_repo.get_remote_url()
         if url is None:
             if targets_relative_dir is not None:
-                url = Path(os.path.relpath(target_repo.repo_path, targets_relative_dir))
+                url = Path(os.path.relpath(target_repo.path, targets_relative_dir))
             else:
-                url = Path(target_repo.repo_path).resolve()
+                url = Path(target_repo.path).resolve()
             # convert to posix path
             url = str(url.as_posix())
         repositories[target_repo_namespaced_name] = {"urls": [url]}
@@ -802,9 +801,10 @@ def register_target_files(
     targets_path = repo_path / TARGETS_DIRECTORY_NAME
     taf_repo = Repository(str(repo_path))
     auth_git_repo = GitRepository(repo_path)
-    target_filenames = []
+
     # find only untracked and modified targets
     if auth_git_repo.is_git_repository:
+        target_filenames = []
         target_files = auth_git_repo.list_modified_files(
             path="targets"
         ) + auth_git_repo.list_untracked_files(path="targets")
@@ -814,18 +814,10 @@ def register_target_files(
                 os.path.relpath(str(modified_file_path), str(targets_path))
             )
     else:
-        for root, _, filenames in os.walk(str(targets_path)):
-            for filename in filenames:
-                filepath = Path(root) / filename
-                if filepath.is_file():
-                    target_filenames.append(
-                        os.path.relpath(str(filepath), str(targets_path))
-                    )
-    targets_roles_mapping = taf_repo.map_signing_roles(target_filenames)
-    for target_rel_path, target_role in targets_roles_mapping.items():
-        taf_repo.add_existing_target(str(targets_path / target_rel_path), target_role)
+        target_filenames = taf_repo.all_target_files()
 
-    updated_targets_roles = set(targets_roles_mapping.values())
+    taf_repo.add_existing_targets(target_filenames)
+    updated_targets_roles = set(taf_repo.map_signing_roles(target_filenames).values())
 
     _write_targets_metadata(
         taf_repo, updated_targets_roles, keystore, roles_key_infos, scheme